refactor: extract cfg-javascript as a library (#103)
