+++ packages/cfg-javascript/NOTICE
@@ -1,0 +1,5 @@
+SynTest Javascript
+Copyright 2020-2023 Delft University of Technology and SynTest contributors
+
+This product includes software developed at
+Delft University of Technology (http://www.tudelft.nl/).
