+++ packages/cfg-javascript/README.md
@@ -1,0 +1,1 @@
+# Control Flow Graph library for JavaScript
