+++ packages/cfg-javascript/lib/CFGGenerator.ts
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2020-2023 Delft University of Technology and SynTest contributors
+ *
+ * This file is part of SynTest Framework - SynTest Javascript.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { traverse } from "@babel/core";
+import { ControlFlowGraphVisitor } from "./ControlFlowGraphVisitor";
+import { CFG } from "@syntest/cfg-core";
+import * as t from "@babel/types";
+
+/**
+ * Exports generator for targets.
+ *
+ * @author Dimitri Stallenberg
+ */
+export class CFGGenerator {
+  /**
+   * Generate function map for specified target.
+   *
+   * @param targetAST The AST of the target
+   */
+  generate(targetAST: t.Node): CFG {
+    const visitor = new ControlFlowGraphVisitor();
+    traverse(targetAST, visitor);
+
+    return visitor.cfg;
+  }
+}
