+++ packages/cfg-javascript/lib/ControlFlowGraphGenerator.ts
