+++ packages/cfg-javascript/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@syntest/javascript",
-  "version": "0.0.2-beta.0",
-  "description": "SynTest JavaScript is a tool for automatically generating test cases for the JavaScript language",
+  "name": "@syntest/cfg-javascript",
+  "version": "0.0.1",
+  "description": "SynTest CFG JavaScript is a library for generating control flow graphs for the JavaScript language",
   "keywords": [
     "syntest",
-    "automatic-test-generation"
+    "control-flow-graph"
   ],
   "homepage": "https://www.syntest.org",
   "bugs": {
@@ -21,14 +21,11 @@
   "files": [
     "/dist",
     "/NOTICE",
-    "/README.md"
+    "README.md"
   ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/syntest-framework/syntest-javascript.git"
-  },
-  "bin": {
-    "syntest-javascript": "/dist/bin.js"
   },
   "scripts": {
     "build": "tsc --build",
@@ -45,23 +42,7 @@
     "test:watch": "mocha --config ../../.mocharc.json --watch"
   },
   "dependencies": {
-    "@babel/core": "7.16.5",
-    "@babel/preset-env": "7.19.1",
-    "@babel/register": "7.18.9",
-    "@istanbuljs/schema": "0.1.3",
-    "@syntest/core": "^0.3.1-beta.0",
-    "@syntest/cfg-javascript": "*",
-    "chai": "4.3.7",
-    "chai-as-promised": "7.1.1",
-    "chalk": "4.1.1",
-    "clear": "0.1.0",
-    "figlet": "1.5.2",
-    "fs-extra": "11.1.0",
-    "istanbul-lib-instrument": "5.1.0",
-    "lodash.clonedeep": "4.5.0",
-    "mocha": "10.2.0",
-    "original-require": "1.0.1",
-    "regenerator-runtime": "0.13.9"
+    "@syntest/cfg-core": "^0.2.1-beta.0"
   },
   "engines": {
     "node": ">=10.24.0"
