+++ packages/cfg-javascript/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../tsconfig-base.json",
+  "files": [],
+  "include": [],
+  "references": [{ "path": "lib/" }]
+}
